reorganizing tests
