+++ crates/opsml_server/Cargo.toml
@@ -48,3 +48,8 @@ mockito = "1.*"
 rand = "0.*"
 serde_qs = "0.*"
 tower = { version = "0.*", features = ["util"] }
+
+
+[[test]]
+name = "integration"
+path = "tests/lib.rs"
