+++ crates/opsml_server/src/lib.rs
@@ -1,4 +1,4 @@
-mod core;
+pub mod core;
 
 pub use core::app::create_app;
 use opsml_colors::Colorize;
