+++ crates/opsml_server/tests/api/mod.rs
@@ -1,0 +1,2 @@
+pub mod card;
+pub mod login;
