+++ crates/opsml_server/tests/lib.rs
@@ -1,0 +1,4 @@
+pub mod api;
+pub mod common;
+
+pub use common::TestHelper;
